Incluindo o FEIGN

+++ book-service/src/main/java/br/com/silber/proxy/CambioProxy.java
@@ -1,7 +1,5 @@
 package br.com.silber.proxy;
 
-import java.math.BigDecimal;
-
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,7 +11,7 @@ public interface CambioProxy {
 	
 	@GetMapping(value = "/cambio-service/{amount}/{from}/{to}")
 	public Cambio getCambio(
-			@PathVariable("amount") BigDecimal amount,
+			@PathVariable("amount") Double amount,
 			@PathVariable("from") String from,
 			@PathVariable("to") String to);
